fix: LLM cost optimization rules UI polish (#1256)
- Removes "optimize cost" toggle from agents / profiles. This means cost
optimization is always on if there is a matching rule in the agent's org
or team.
- All rules are grouped by models and evaluated with && i.e. `max 1000
tokens && without tools`. Otherwise if they are separate rules, requests
without tools but with 99999 tokens will try to use the cheaper model
and will fail. We may go for fancy compound rule UI in the future, but
for now this sounds sufficient.

Closes https://github.com/archestra-ai/archestra/issues/1105